buffer_pool

Struct Pool

pub struct Pool<const S: usize, T: 'static> { /* private fields */ }

A sharded pool of elements.

Implementations

impl<T: Default + Reuse> Pool<{ 32 }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<T: Default + Reuse> Pool<{ _ }, T>

pub const fn new(limit: usize, trim: usize) -> Self

impl<const S: usize, T: Default + Reuse> Pool<S, T>

pub fn get(&'static self) -> Pooled<T>

Get a value from the pool, or create a new default value if the assigned shard is currently empty.

pub fn get_empty(&'static self) -> Pooled<T>

Create a new default value assigned for a pool, if it is ends up being expanded and eligible for reuse it will return to the pool, otherwise it will end up being dropped.

pub fn get_with(&'static self, f: impl Fn(&mut T)) -> Pooled<T>

Get a value from the pool and apply the provided transformation on it before returning.

pub fn from_owned(&'static self, inner: T) -> Pooled<T>

impl<'a, const S: usize, T: Default + Extend<&'a u8> + Reuse> Pool<S, T>

pub fn with_slice(&'static self, v: &'a [u8]) -> Pooled<T>

Get a value from the pool and extend it with the provided slice.