buffer_pool

Trait Reuse

pub trait Reuse {
    // Required methods
    fn reuse(&mut self, trim: usize) -> bool;
    fn capacity(&self) -> usize;
}

A trait that prepares an item to be returned to the pool. For example clearing it. true is returned if the item should be returned to the pool, false if it should be dropped.

Required Methods

fn reuse(&mut self, trim: usize) -> bool

fn capacity(&self) -> usize

Returns the capacity of the object in bytes, to allow for more precise tracking.