buffer_pool

Trait Reuse

pub trait Reuse {
    // Required method
    fn reuse(&mut self, trim: usize) -> bool;
}

A trait that prepares an item to be returned to the pool. For example clearing it. true is returned if the item should be returned to the pool, false if it should be dropped.

Required Methods

fn reuse(&mut self, trim: usize) -> bool